upgrades working (sort of)
Everything for upgrades works, purchase locked if you dont have enough bubbles.
there is a colour change test to see when they are upgraded and the upgrade panel can be opened/closed. Only issue is that the text doesnt properly update for the upgrade costs this may be to do with how i retrieve the cost integer from the resource manager.